SeriesIdInput refactoring part#2
